Refactor : FruitStorage, JuiceRecipes 파일 분리
